Reorganization and clean up imports
