+++ src/backends/docx/mod.rs
@@ -1,4 +1,33 @@
 //! This module contains document and style-building code for the docx output
 pub mod document;
-pub mod styles;
 pub mod numbering;
+pub mod styles;
+
+use std::{fs::File, path::Path};
+
+use docx_rs::{DocxError, Paragraph};
+
+use crate::graph::asg::Asg;
+
+/// !Experimental! Renders a Docx file. Some [`Asg`] blocks are still unsupported.
+pub fn render_docx(graph: &Asg, output_path: &Path) -> Result<(), DocxError> {
+    let file = File::create(output_path).unwrap();
+    let mut writer = document::DocxWriter::new();
+    let mut docx = document::asciidocr_default_docx();
+
+    // Add document title if present
+    if let Some(header) = &graph.header {
+        if !header.title.is_empty() {
+            let mut para = Paragraph::new().style("Title");
+            para = document::add_inlines_to_para(para, header.title());
+            docx = docx.add_paragraph(para);
+        }
+    }
+
+    // Add document contents
+    for block in graph.blocks.iter() {
+        docx = writer.add_block_to_doc(docx, block)
+    }
+    docx.build().pack(file)?;
+    Ok(())
+}
